nodediscovery: split StartDiscovery into smaller functions
in preparation to a subsequent change, this commit moves some of the
(*NodeDiscovery).StartDiscovery() logic into separated functions:

* fillLocalNode(), responsible for syncing the localNode object with the
  actual state of the node
* updateLocalNode(), responsible for publishing the updated KV store
  entry and/or CiliumNode object for the local node

In addition to that, this commit introduces also the UpdateLocalNode()
function, which ensure that the proper lock is held while calling
updateLocalNode(), and allows optionally to sync the localNode object
state by calling fillLocalNode().

No functional changes are introduced.

Signed-off-by: Gilberto Bertin <jibi@cilium.io>